requirements.txt
boto3
pandas
duckdb
streamlit
pyarrow
python-dotenv
